requirements.txt
aiogram==3.13.0
python-dotenv==1.0.0
openai==1.12.0
aiohttp==3.9.1
python-json-logger==2.0.7
